+++ public/index.php
@@ -1,7 +1,24 @@
 <?php
 
-require dirname(__DIR__) . '/vendor/autoload.php';
+declare(strict_types=1);
 
 use App\Application;
+use DI\ContainerBuilder;
+use Illuminate\Database\Capsule\Manager as CapsuleManager;
 
-$app = new Application();
+require dirname(__DIR__) . '/vendor/autoload.php';
+
+$builder = new ContainerBuilder();
+
+$builder->useAutowiring(true);
+
+$builder->addDefinitions(
+    dirname(__DIR__) . '/config/container.php'
+);
+
+$container = $builder->build();
+
+$container->get(CapsuleManager::class);
+
+$application = $container->get(Application::class);
+$application->run();